third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/img-onerror-non-ASCII-char-default-encoding-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag-default-encode.pl?q=%3Cimg+src=%27%80%27+onerror=%27alert(document.domain)%27' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
